api export: Dynamic export of symbols function
