+++ plugins/geanyfunctions.h
@@ -16,15 +16,9 @@
 #ifndef GEANY_FUNCTIONS_H
 #define GEANY_FUNCTIONS_H
 
-#define build_activate_menu_item \
-	geany_functions->p_build->build_activate_menu_item
-#define build_get_current_menu_item \
-	geany_functions->p_build->build_get_current_menu_item
-#define build_remove_menu_item \
-	geany_functions->p_build->build_remove_menu_item
-#define build_set_menu_item \
-	geany_functions->p_build->build_set_menu_item
-#define build_get_group_count \
-	geany_functions->p_build->build_get_group_count
+/* This header is deprecated. use geanyplugin.h directly */
+#ifndef GEANY_DISABLE_DEPRECATED
+#include "geanyplugin.h"
+#endif
 
 #endif
