+++ plugins/genapi.py
@@ -64,6 +64,7 @@ def get_function_names():
                     or symbol.startswith('main_')
                     or symbol.startswith('stash_')
                     or symbol.startswith('symbols_')
+                    or symbol.startswith('build_')
                     or symbol.endswith('_funcs')):
                 names.append(symbol)
     filep.close()
